requirements.txt
certifi==2018.8.24
dj-database-url==0.5.0
Django==2.1.2
django-heroku==0.3.1
django-rest-multiple-models==2.1.1
django-shortuuidfield==0.1.3
djangorestframework==3.8.2
gunicorn==19.9.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
pytz==2018.5
shortuuid==0.5.0
six==1.11.0
virtualenv==16.0.0
virtualenv-clone==0.3.0
whitenoise==4.0
